+++ src/integrations/prefect-ray/tests/test_task_runners.py
@@ -13,7 +13,7 @@ from prefect_ray.context import remote_options
 
 import prefect
 import prefect.engine
-import prefect.new_task_engine
+import prefect.task_engine
 import tests
 from prefect import flow, task
 from prefect.states import State, StateType
