+++ tests/test_task_engine.py
@@ -20,13 +20,13 @@ from prefect.context import (
 )
 from prefect.exceptions import CrashedRun, MissingResult
 from prefect.filesystems import LocalFileSystem
-from prefect.new_task_engine import TaskRunEngine, run_task_async, run_task_sync
 from prefect.results import PersistedResult, ResultFactory
 from prefect.settings import (
     PREFECT_TASK_DEFAULT_RETRIES,
     temporary_settings,
 )
 from prefect.states import Running, State
+from prefect.task_engine import TaskRunEngine, run_task_async, run_task_sync
 from prefect.task_runners import ThreadPoolTaskRunner
 from prefect.testing.utilities import exceptions_equal
 from prefect.utilities.callables import get_call_parameters
